Assorted Changes (CSS + API + Old Stuff)
- Updated few css styles for animation
- Updated homepage slightly
- Updated API calls to beter support eactivities sync
- Updated API call for locker sales (work-in-progress)
- Added unsynced changes from last year
- Added new and older (unsynced) db migrations

+++ app/controllers/DinnerGroupsController.php
@@ -2,13 +2,16 @@
 
 class DinnerGroupsController extends BaseController {
 
-    const MAX_TABLES = 19;
+    const MAX_TABLES = 24;
 
-    protected $hasExpired = TRUE;
+    protected $hasExpired = FALSE; /* (time() >= strtotime("2017-01-13 23:59")); */
+
 
     public function __construct()
     {
         parent::__construct();
+
+        $this->hasExpired = time() >= strtotime("2017-01-14 23:59");
 
         $this->beforeFilter(function() {
             if (!DinnerPermission::user(Auth::user())->canManageGroups()) {
@@ -18,7 +21,7 @@ class DinnerGroupsController extends BaseController {
 
             if ($this->hasExpired && !Auth::user()->is_admin) {
                 return Redirect::action('UsersController@getDashboard')
-                    ->with('danger', 'The period for choosing your seat has ended. Should you require further assistance, please <a href="mailto:eesoc.webmaster@imperial.ac.uk">email us</a>.');
+                    ->with('danger', 'The period for choosing your seat has ended. Should you require further assistance, please <a href="mailto:eesoc.events@imperial.ac.uk">email us</a>.');
             }
         });
     }
@@ -32,7 +35,7 @@ class DinnerGroupsController extends BaseController {
     {
         return View::make('dinner_groups.index')
             ->with('user', Auth::user())
-            ->with('groups', DinnerGroup::all());
+            ->with('groups', DinnerGroup::orderBy('id')->get());
     }
 
     /**
@@ -60,6 +63,8 @@ class DinnerGroupsController extends BaseController {
                                        .'If you wish to join another group, '
                                        .'please leave this one.');
         }
+
+        $user->getUnclaimedDinnerTicketsCountAttribute();
 
         // Show multiple user info form if multiple tickets owned to allocate.
         if ($user->unclaimed_dinner_tickets_count > 1) {
@@ -99,6 +104,8 @@ class DinnerGroupsController extends BaseController {
                          .'If you wish to join another group, '
                          .'please leave this one.');
         }
+
+        $user->getUnclaimedDinnerTicketsCountAttribute();
 
         if ($user->unclaimed_dinner_tickets_count === 1)
             return Redirect::route('dashboard.dinner.groups.create');
@@ -188,10 +195,14 @@ class DinnerGroupsController extends BaseController {
                 ->with('danger', 'You cannot leave this group');
         }
 
-        $group->removeMember($user);
+        $removed = $group->removeMember($user);
 
-        return Redirect::route('dashboard.dinner.groups.show', $group->id)
-            ->with('success', 'You have left this group');
+        if ($removed === true) {
+            return Redirect::route('dashboard.dinner.groups.show', $group->id)
+                ->with('success', 'You have left this group');
+        }else{
+            return $removed;
+        }
     }
 
     public function removeMember()
@@ -206,7 +217,29 @@ class DinnerGroupsController extends BaseController {
                 ->with('danger', 'You cannot remove that guest from their table.');
         }
 
+        if ($member->is_owner && !DinnerGroup::CAN_LEAVE_OWN_GRP)
+        {
+            return Redirect::route('dashboard.dinner.groups.show', $oldGroup)
+                ->with('danger', 'You cannot leave from your own group.');
+        }
+
+
+        $group = $member->dinnerGroup()->first();
+        //$membersInGroup = $group->members()->get()->count();
+        $membersInGroup = DinnerGroupMember::where('dinner_group_id','=',$oldGroup)->count();
+
         $member->delete();
+
+        //If user was the last person to leave the group, delete the group as well.
+        if ($membersInGroup <= 1){
+            $group->delete();
+            return Redirect::route('dashboard.dinner.groups.index')
+                ->with('success', 'You have left the group.');
+        }
+		
+		$group->is_full = $membersInGroup - 1 >= $group->max_size;
+        
+		$group->save();		
 
         return Redirect::route('dashboard.dinner.groups.show', $oldGroup)
             ->with('success', 'Your guest was removed.');
@@ -217,6 +250,8 @@ class DinnerGroupsController extends BaseController {
         $user     = Auth::user();
         $group    = DinnerGroup::findOrFail((integer) Input::get('group'));
         $oldGroup = $group->id;
+
+        $user->getUnclaimedDinnerTicketsCountAttribute();
 
         if (!DinnerPermission::user($user)->canAddUserToGroup($group) ||
             (!$user->DinnerGroupMember && $user->unclaimed_dinner_tickets_count <= 1) ||
@@ -234,38 +269,43 @@ class DinnerGroupsController extends BaseController {
     public function updateMenuChoice()
     {
         $member = DinnerGroupMember::findOrFail((integer) Input::get('member'));
-        $course = Input::get('course');
-        $choice = Input::get('choice');
+        $courses = [
+            "starter",
+            "main",
+            "dessert"
+        ];
 
-        switch ($course)
-        {
-        case 'starter':
-        case 'main':
-            $vName = "vegetarian_$course";
-            break;
+        foreach($courses as $course){
+            $vName = "choice_$course";
+            $choice = Input::get($vName);
 
-        default:
-            throw new \InvalidArgumentException("An invalid course type was specified.");
+            echo $vName." -> ".$choice;
+            //Choice validation
+            switch ($course)
+            {
+            case 'main':
+                if ($choice <= 4 && $choice >= 1){ break; }
+
+            case 'starter':
+            case 'dessert':
+                if ($choice <= 3 && $choice >= 1){ break; }
+
+            default:
+                return Redirect::route('dashboard.dinner.groups.show', $member->dinner_group->id)
+                    ->with('danger', 'Please select the menu choices fo all courses for '.$member->name.'.');
+                //throw new \InvalidArgumentException("An invalid choice was specified.");
+
+            }
+
+            $member->$vName = $choice;
         }
 
-        switch ($choice)
-        {
-        case 'meat':
-            $vegetarian = FALSE;
-            break;
 
-        case 'vegetarian':
-            $vegetarian = TRUE;
-            break;
+        $member->special_req = Input::get('special_req');
 
-        default:
-            throw new \InvalidArgumentException("An invalid choice was specified.");
-        }
-
-        $member->$vName = $vegetarian;
         $member->save();
 
         return Redirect::route('dashboard.dinner.groups.show', $member->dinner_group->id)
-            ->with('success', 'Menu choice updated.');
+            ->with('success', 'Thanks! The menu choice updated for '.$member->name.'.');
     }
 }
